requirements.txt
streamlit==1.24.0
SQLAlchemy==2.0.25
langchain==0.1.4
langchain-community==0.0.17
langchain-core==0.1.18
langchain-experimental==0.0.42
langchain-openai==0.0.5
openai==1.10.0
pdf2image==1.17.0
pdfminer.six==20221105
pinecone-client==3.0.2
pypdf==4.0.1
PyPDF2==3.0.1
python-dotenv
protobuf
sentence-transformers